cuify/Server/requirements.txt
boto3
openai
protobuf
python-dotenv
PyYAML
scipy
transformers
google-generativeai
torch
ffmpeg
pydub
accelerate
huggingface_hub